fix: centralize writeError signature
